Fixed type

+++ localization/nerf_based_localizer/src/nerf/common_cuda.hpp
@@ -7,17 +7,17 @@
 
 #include <cuda_runtime.h>
 
-inline int DivUp(const int x, const int y)
+inline unsigned int DivUp(const unsigned int x, const unsigned int y)
 {
   return (x + y - 1) / y;
 }
 
-constexpr int THREAD_CAP = 512;
-constexpr dim3 LIN_BLOCK_DIM = {unsigned(THREAD_CAP), 1, 1};
+constexpr unsigned int THREAD_CAP = 512;
+constexpr dim3 LIN_BLOCK_DIM = {THREAD_CAP, 1, 1};
 
 inline dim3 LIN_GRID_DIM(const int x)
 {
-  return dim3{unsigned(DivUp(x, THREAD_CAP)), 1, 1};
+  return dim3{DivUp(x, THREAD_CAP), 1, 1};
 }
 
 #endif  // NERF_BASED_LOCALIZER__COMMON_CUDA_HPP_
